requirements.txt
discord.py==1.3.3
youtube-dl==2020.3.24
PyNaCl==1.3.0
spotdl==1.2.6
spotipy==2.11.1
numpy==1.18.2
